refactor: extract inline test blocks to dedicated tests modules
Move the inline #[cfg(test)] blocks out of two oversized files into
dedicated test submodules, matching the monolith-refactor convention.

- zeph-subagent/src/manager/mod.rs (4118 -> 527 lines): the three
  top-level test modules move to manager/tests.rs; the main `tests`
  module is de-indented to file top level, while worktree_predicate_tests
  and worktree_cleanup_guard_tests become submodules of tests.
- zeph-orchestration/src/scheduler/tick.rs (2200 lines) becomes a
  directory: tick/mod.rs holds production code, tick/tests.rs the tests.

Production code is byte-identical in both files; pure mechanical move
with no behavior change. All 159 moved tests (113 + 46) still pass.

Closes #4924
